+++ src/com/ignatenko/figures/Shape.java
@@ -2,6 +2,6 @@ package com.ignatenko.figures;
 
 public interface Shape {
 
-	public double getArea();
+	public int getArea();
 
 }
